test(red): 출석 날짜가 주말이면 예외 발생 테스트 추가
